# docs/swagger.yaml
openapi: 3.0.2
info:
    title: games-api
    version: 1.0.0
    description: A brand new API with no content.  Go nuts!
paths:
    /games:
        get:
            requestBody:
                content:
                    new-game: {}
                required: true
            tags:
                - games
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    type: string
                            examples:
                                List of games:
                                    value: |-
                                        [
                                            'Witcher 3',
                                            'Cyberpunk 2077'
                                        ]
                    description: A list of games
                '201':
                    content:
                        new-game: {}
                    description: '123'
                '400':
                    description: One or all params are missing
            summary: Gets a list of games
            description: Returns a list of games
    '/games/{id}':
        get:
            tags:
                - details
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    type: string
                            examples:
                                Full details of a given game:
                                    value: |-
                                        [
                                            {id: 1, name: "Witcher 3", price: 29.99},
                                            {id: 2, name: "Cyberpunk 2077", price: 59.99},
                                            {id: 3, name: "Minecraft", price: 26.99},
                                            {id: 4, name: "Counter-Strike: Global Offensive", price: 0},
                                            {id: 5, name: "Roblox", price: 0},
                                            {id: 6, name: "Grand Theft Auto V", price: 29.99},
                                            {id: 7, name: "Valorant", price: 0},
                                            {id: 8, name: "Forza Horizon 5", price: 59.99}
                                        ]
                    description: Full details of a given game
                '404':
                    content:
                        application/json:
                            schema:
                                type: string
                            examples:
                                Error message:
                                    value:
                                        error: Error message.
                    description: Return error when no game
            operationId: id
            summary: Get full details of a game
            description: Returns full details of a given id game
        delete:
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/delete-game'
            responses:
                '204': {}
                '404': {}
components:
    schemas:
        delete-game:
            description: ''
            required:
                - id
            type: object
            properties:
                id:
                    description: ''
                    type: string
            example:
                id: 1
        new-game:
            title: Root Type for new-game
            description: ''
            required:
                - price
                - name
            type: object
            properties:
                name:
                    type: string
                price:
                    format: float
                    type: number
            example:
                name: Tetris
                price: 0.99